Move IP checking to common utils
